+++ spec/ultra_settings/configuration_view_spec.rb
@@ -42,10 +42,10 @@ RSpec.describe UltraSettings::ConfigurationView do
       expect(html).not_to include("<a href=")
     end
 
-    it "renders links for the runtime settings" do
+    it "renders links for the runtime settings", settings: {} do
       UltraSettings.runtime_settings_url = "http://example.com/settings?filter=${name}"
       html = UltraSettings::ConfigurationView.new(TestConfiguration.instance).render
-      expect(html).to include('<a href="http://example.com/settings?filter=test.string"')
+      expect(html).to include('href="http://example.com/settings?filter=test.string"')
     end
   end
 end