+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.8.0
+
+### Added
+
+- Added support for setting a description on configuration classes. The class description can serve to document the purpose of the configuration and will be shown in the web UI.
+- Improved menu for selecting configuration classes in the web UI by adding a search box to filter the list of configurations.
+
+### Fixed
+
+- Fixed filtering of data sources for configuration classes in the web UI.
+  - The YAML source will not be shown if the configuration file is set to nil or false.
+  - The runtime settings source will not be shown if the runtime settings engine is not set up.
+  - Fields that override the class default for a data source will now correctly show that source.
+
 ## 2.7.0
 
 ### Added